custom Authentication

+++ src/main/java/com/ecommerce/project/model/Role.java
@@ -15,7 +15,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @ToString.Exclude
     @Enumerated(EnumType.STRING)
@@ -25,4 +25,5 @@ public class Role {
     public Role(AppRole roleName) {
         this.roleName = roleName;
     }
+
 }
